Input | Floor Material
Keyboard input is now instant. Prior to this change it would take a number of frames to go from 0->1 and 1->0 when the player pressed and released the button. This is meant to mimic the analog sticks but feels wrong in a 2D game, and causes the player to slide when he shouldn't.

Also added a new physics material called Floor. Now actually tuned in any way yet, and should not be used.