Jump Pad
Fixing jump pads to work with the player. Jump pads now need to be tagged with "JumpPad", but that is part of the prefab.

The Player Controller has been altered a bit to have both a "CurrentVelocity" and "WantedVelocity". 'Wanted' is what methods will alter during the Update pass, but then Current can be used in the fixed update to make changes over time.