Spawner
Spawner can now be killed. When it is killed, it plays the Open animation and stays like that. It also fades to a grayed-out tint.

The way damage flash was working was causing it to override the color over and over when it wasn't active, negating any other Behaviours that wanted to change the color. That has been fixed.