Player Controller
Player will no longer switch to jump animation when walking down steep slopes. To do this we set a min amount of time he has to be in the air before switching to jump (when falling) and also do a ray cast check. Both have to pass before the walk animation is used.