Flyer AI
Fixed issue with flyer not dying. This was because the new animation did not have multiple frame and the scripted was waiting for an animation delegate to get called before deleting the object. Fly now just uses the jumper animation as placeholder.

Note: This change delete Resources.meta but I am not sure what that file is. I suspect Erik has something not checked in.