Launcher Weapon (WIP)
WIP on Launcher Weapon. Basic shot is in, but most of the changes here were to make the existing weapon systems a little more generic to support more variety in the weapons.

Also made it so that weapons can be fired as fast as the player can press the button so that it always feels responsive, especially on slower firing weapons.