Blaster Weapon
Gave Blaster weapon a bit of randomness to its firing angles to make it feel a little more out of control and powerful. I wanted the bullet to be rotated to face their firing direction, so that meant getting rid of the existing code which flipped the bullet when fired.