Level Loading
Fixing issues with levels being loaded and appearing offset from original position in running version of the game. This was a bug related to the Camera having a height of 8 but then at runtime changing that height to 9. The level loading script uses that height to determind the position of tiles and collision.

Also changes the Tilemap loading back to the original version where the collision mesh is loaded as a singlemesh since we aren't using physics materials any more.