Player State Save Between Levels
The player is no longer destroyed and recreated between levels. This means he now keeps his equiped weapon and health is not restored.

This required some changes to the way the player is spawned. You no longer need to support both placing him in the level and spawning from the Game Instance; the Instance now always handles it, even when not streaming in a level.